fixed some bugs in GameBoardPanel

+++ src/main/java/it/unibo/monoopoly/view/panel/impl/Position.java
@@ -1,3 +1,3 @@
 package it.unibo.monoopoly.view.panel.impl;
 
-public record Position(int x, int y) {}
+public record Position(double x, double y) {}
